==> 1_cervejas.ipynb <==
### FAZ A EXTRAÇÃO DOS ESTILOS DE CERVEJA DO JSON DO BJCP PARA UM CSV ###

import requests
import pandas as pd
import numpy as np

# URL para JSON BJCP - Estilos de cervejas artesanais e atributos

url = "https://raw.githubusercontent.com/ascholer/bjcp-styleview/refs/heads/main/styles.json"


# FAZ REQUISIÇÃO E CRIA DATAFRAME

# Fazer a requisição e carregar os dados JSON
response = requests.get(url)
json_data = response.json()

# Criar um DataFrame a partir dos dados JSON
beer_styles_df = pd.json_normalize(json_data)

# Exibir o DataFrame
print(beer_styles_df.head(10))

# REMOVE AS LINHAS COM VALORES NULOS

# foi selecionado apenas os atributos importantes e que portanto, que não podem ter valores nulos
beer_styles_df = beer_styles_df.dropna(subset=['name', 'flavor', 'ibumin', 'ibumax', 'ogmin', 'ogmax', 'fgmin', 'fgmax', 'abvmin', 'abvmax', 'srmmin', 'srmmax'])

# Exibe o DataFrame limpo
beer_styles_df.head(10)

# SALVANDO RESULTADO EM CSV

beer_styles_df.to_csv("./datasets/raw/CERVEJAS.csv", index=False)


# verificando a quantidade de estilos de cervejas
qtdd_beer_styles = beer_styles_df.shape[0]
print(qtdd_beer_styles)

df_teste = pd.read_csv('./datasets/raw/CERVEJAS.csv')

# colunas disponiveis
print(df_teste.columns)


df_teste[['name', 'flavor', 'mouthfeel', 'ibumin', 'ibumax',
       'ogmin', 'ogmax', 'fgmin', 'fgmax', 'abvmin', 'abvmax', 'srmmin',
       'srmmax']].to_csv("./testeCerva.csv", index=False)
