==> 2_pratos_harmonizacao.ipynb <==
### EXTRAI APENAS OS NOMES DOS PRATOS DA BASE DE HARMONIZACAO ###

import pandas as pd

# LEITURA DO ARQUIVO DE HARMONIZACOES, QUE CONTEM OS PRATOS

harmonizacao_df = pd.read_csv('./datasets/preprocessed/HARMONIZACOES_preprocessed.csv')

# EXTRAI PRATOS DA BASE DE HARMONIZAÇÕES

# Criar um conjunto para armazenar preparações únicas
pratos = set()

# Iterar pelas linhas do DataFrame para coletar preparações
for preparacoes in harmonizacao_df['pratos']:
    # Verifica se preparacoes é um conjunto em formato de string e converte para um conjunto
    if isinstance(preparacoes, str):
        preparacoes_set = eval(preparacoes)  # Converte a string do conjunto de volta para um conjunto
    else:
        preparacoes_set = preparacoes  # Caso já seja um conjunto

    # Adicionar as preparações ao conjunto de preparações distintas
    pratos.update(preparacoes_set)

# Criar uma lista de preparações distintas
lista_pratos = list(pratos)

# Traduzir as preparações distintas
#lista_pratos_en = [translator.translate(preparacao, destination_language='en').result for preparacao in lista_pratos]

# Criar um DataFrame com as colunas 'dishes' e 'dishes_translated'
pratos_df = pd.DataFrame({
    'pratos': lista_pratos
})

# Exportar o DataFrame para um arquivo CSV
pratos_df.to_csv("./datasets/preprocessed/PRATOS.csv", index=False)

# Exibir as primeiras linhas do DataFrame final
print(pratos_df.head())
