==> 3_pratos_ingredientes.ipynb <==
### SEPARAÇÃO DA BASE DE PRATOS DE TREINAMENTO E A DE PREDICÃO ###

### OS PRATOS DA BASE ALLRECIPES QUE DAO MATCH COM A BASE DE PRATOS QUE VEIO DA LISTA DE HARMONIZACAO
### SERAO OS PRATOS USADOS PARA TREINAMENTO

### O RESTANTE SERÁ USADO COMO TESTE DE PREDIÇÃO COM O ALGORITMO PRONTO


import pandas as pd
from fuzzywuzzy import process


# Carregar as bases de dados
pratos_allrecipes_filtrados = pd.read_csv('./datasets/raw/ALLRECIPES_PRATOSxINGREDIENTES_filtrado.csv')
pratos_harmonizacao = pd.read_csv('./datasets/preprocessed/PRATOS.csv')

# separando as receitas que tem na minha base de PRATOS das que nao tem

# Listas de nomes de receitas
allrecipe_names = pratos_allrecipes_filtrados['prato'].tolist()
pratos_harmonizacao_names = pratos_harmonizacao['pratos'].tolist()

# Tolerância de similaridade (0 a 100), onde 100 é uma correspondência exata
similarity_threshold = 90

# Encontrar receitas similares
def find_pratos_similares(recipes, pratos, threshold):
    matches = []
    for recipe in recipes:
        best_match, score = process.extractOne(recipe, pratos)
        if score >= threshold:
            matches.append((recipe, best_match, score))
    return matches

# Executar a comparação
pratos_similares = find_pratos_similares(allrecipe_names, pratos_harmonizacao_names, similarity_threshold)

# Criar dataframes para as novas bases
similarity_df = pd.DataFrame(pratos_similares, columns=['pratos_allrecipes', 'pratos_harmonizacao', 'similarity_score'])

# Identificar receitas que estão na base de pratos e as que não estão
pratos_match = similarity_df['pratos_allrecipes'].unique()
pratos_novasEntradas = pratos_allrecipes_filtrados[~pratos_allrecipes_filtrados['prato'].isin(pratos_match)]
pratos_treinamento = pratos_allrecipes_filtrados[pratos_allrecipes_filtrados['prato'].isin(pratos_match)]



# Fazer o merge da base de treinamento com similarity_df para adicionar a coluna pratos_harmonizacao
pratos_treinamento = pd.merge(pratos_treinamento, similarity_df[['pratos_allrecipes', 'pratos_harmonizacao']], 
                        left_on='prato', right_on='pratos_allrecipes', how='left')

print(pratos_treinamento)


# BASE QUE TERÁ TODOS OS MATCHINGS DE PRATOS E INGREDIENTES

df_pratos_ingredientes = pd.concat([pratos_novasEntradas, pratos_treinamento], ignore_index=True)


print(df_pratos_ingredientes)

# remove duplicadas
df_pratos_ingredientes = df_pratos_ingredientes.drop_duplicates()


df_pratos_ingredientes.to_csv('./datasets/preprocessed/PRATOSxINGREDIENTES.csv', index=False)


# Manter apenas as colunas 'prato' (que vem da base harmonizacao), prato_allrecipes e 'ingredientes'
pratos_treinamento = pratos_treinamento.drop(columns=['prato'])

pratos_treinamento = pratos_treinamento.rename(columns={'pratos_harmonizacao': 'prato'})


# remove duplicadas
pratos_treinamento = pratos_treinamento.drop_duplicates()


pratos_treinamento

# Salvar as novas bases em arquivos CSV, se necessário
pratos_novasEntradas.to_csv('./datasets/final/PRATOSxINGREDIENTES_novasEntradas.csv', index=False)
pratos_treinamento.to_csv('./datasets/train/PRATOSxINGREDIENTES_treinamento.csv', index=False)

# Salvar o DataFrame de similaridade, se necessário
#similarity_df.to_csv('similarity_df.csv', index=False)
