==> 4_tratamento_base_cerveja.ipynb <==
import pandas as pd
from sklearn.preprocessing import MinMaxScaler
import numpy as np
from itertools import product

# carregar os datasets
cervejas = pd.read_csv('./datasets/raw/CERVEJAS.csv')
harmonizacoes = pd.read_csv('./datasets/preprocessed/HARMONIZACOES_processed.csv')

cervejas[['name','ibumin', 'ibumax', 'ogmin', 'ogmax', 'fgmin', 'fgmax', 'abvmin', 'abvmax', 'srmmin', 'srmmax']]

# criacao de faixas entre os valores de min e max de cada um dos atributos para 
# eliminar as colunas de min e max, ficando apenas 1 coluna por atributo

def criar_faixas(df, num_faixas=3):
    faixas = []
    for _, row in df.iterrows():
        for i in range(num_faixas):
            faixa = {
                'estilo': row['name'],
                'categoria': row['category'],
                'ibu': np.linspace(row['ibumin'], row['ibumax'], num_faixas)[i],
                'og': np.linspace(row['ogmin'], row['ogmax'], num_faixas)[i],
                'fg': np.linspace(row['fgmin'], row['fgmax'], num_faixas)[i],
                'abv': np.linspace(row['abvmin'], row['abvmax'], num_faixas)[i],
                'srm': np.linspace(row['srmmin'], row['srmmax'], num_faixas)[i]
            }
            faixas.append(faixa)
    return pd.DataFrame(faixas)

# criando faixas de parâmetros com valores originais
cervejas = criar_faixas(cervejas)

cervejas[['estilo','ibu', 'og', 'fg', 'abv', 'srm']]

# Definir as colunas dos atributos das cervejas que queremos normalizar
range_atributos_cerveja = ['ibumin', 'ibumax', 'ogmin', 'ogmax', 'fgmin', 'fgmax', 'abvmin', 'abvmax', 'srmmin', 'srmmax']
atributos_cerveja = ['ibu', 'og', 'fg', 'abv', 'srm']

# gerando todas as possíveis combinacoes entre os valores de faixa para os atributos
# cada linha gerada simula um possível rótulo de cerveja para um estilo

def gerar_combinacoes(df_faixas):
    faixas_combinadas = []
    for estilo in df_faixas['estilo'].unique():
        sub_df = df_faixas[df_faixas['estilo'] == estilo]
        combinacoes = product(sub_df['ibu'], sub_df['og'], sub_df['fg'], sub_df['abv'], sub_df['srm'])
        
        for combo in combinacoes:
            faixa_comb = {
                'estilo': estilo,
                'ibu': combo[0],
                'og': combo[1],
                'fg': combo[2],
                'abv': combo[3],
                'srm': combo[4]
            }
            for col in sub_df.columns:
                if col not in faixa_comb:
                    faixa_comb[col] = sub_df[col].values[0]  # Acessa o valor correspondente ao estilo
            faixas_combinadas.append(faixa_comb)
    
    return pd.DataFrame(faixas_combinadas)

# gerando as combinações a partir das faixas
cervejas = gerar_combinacoes(cervejas)

cervejas_altbier = cervejas[cervejas['estilo'] == 'Altbier'][['estilo', 'ibu', 'og', 'fg', 'abv', 'srm']]
cervejas_altbier

# verificando balanceamento por estilo e por categoria

# Contagem de linhas por categoria (verificação) e ordenação
row_count_per_estilo = cervejas['estilo'].value_counts().sort_values(ascending=False)
row_count_per_category = cervejas['categoria'].value_counts().sort_values(ascending=False)

# Contagem de estilos únicos por categoria (como antes) e ordenação
# unique_styles_per_category = df_novo_harmonizacao.groupby('category')['estilos_cervejas'].nunique().sort_values(ascending=False)

# Exibir os resultados
print("Número total de linhas por categoria (ordenado do maior para o menor):\n", row_count_per_estilo)
print("\nNúmero de estilos de cervejas únicos por categoria (ordenado do maior para o menor):\n", row_count_per_category)


# normalizando os valores para todos se encaixarem entre 0 e 1, respeitando a proporção

# inicializando o normalizador MinMaxScaler
scaler = MinMaxScaler()

cervejas[atributos_cerveja] = scaler.fit_transform(cervejas[atributos_cerveja])

cervejas_altbier = cervejas[cervejas['estilo'] == 'Altbier'][['estilo', 'ibu', 'og', 'fg', 'abv', 'srm']]
cervejas_altbier

# exportando base tratada de cerveja

cervejas.to_csv('./datasets/preprocessed/CERVEJAS_processed.csv', index=False)
