==> 4_tratamento_base_ingrediente.ipynb <==
import pandas as pd
from collections import Counter
from sklearn.preprocessing import MinMaxScaler



# carregando o dataset
ingredientes = pd.read_csv('./datasets/preprocessed/INGREDIENTES_preprocessed.csv')


ingredientes_chocolate = ingredientes[ingredientes['ingrediente'] == 'chocolate']

ingredientes_chocolate

ingredientes

# função para expandir os sabores e pesos

def expand_peso_sabores(peso_sabores):
    # avaliar o Counter para obter o dicionário de pesos
    return eval(peso_sabores)
    

# aplicar a função para expandir a coluna 'peso_sabores'
ingredientes_expandido = ingredientes['peso_sabores'].apply(expand_peso_sabores)


# criando um DataFrame a partir dos dicionários de pesos de sabores
ingredientes_sabores_df = pd.DataFrame(ingredientes_expandido.tolist())


# preenchendo valores NaN com 0 (ingredientes que não têm aquele sabor)
ingredientes_sabores_df = ingredientes_sabores_df.fillna(0)



# anexando as colunas 'ingrediente' e 'categoria' ao novo DataFrame
ingredientes_sabores_df = pd.concat([ingredientes[['ingrediente', 'categoria']], ingredientes_sabores_df], axis=1)



ingredientes_sabores_df[['ingrediente','categoria','sour', 'bitter', 'sweet', 'fruity', 'grassy','roast', 'caramel', 'fermented','powdery','dry']]

# salvando a base tratada em csv
ingredientes_sabores_df.to_csv('./datasets/preprocessed/INGREDIENTES_processed.csv', index=False)


todos_sabores = [col for col in ingredientes_sabores_df.columns if col not in ['ingrediente', 'categoria']]


# normalizando os valores para todos se encaixarem entre 0 e 1, respeitando a proporção

ingredientes_sabores_norm_df = ingredientes_sabores_df.copy()

scaler = MinMaxScaler()

ingredientes_sabores_norm_df[todos_sabores] = scaler.fit_transform(ingredientes_sabores_norm_df[todos_sabores])

ingredientes_sabores_norm_df[['ingrediente','categoria','sour', 'bitter', 'sweet', 'fruity', 'grassy','roast', 'caramel', 'fermented','powdery','dry']]

# salvando a base tratada e normalizada em csv
ingredientes_sabores_norm_df.to_csv('./datasets/preprocessed/INGREDIENTES_normalizados_processed.csv', index=False)
