==> 5_tratamento_base_prato.ipynb <==
import pandas as pd
from sklearn.preprocessing import MinMaxScaler


# Carregar os datasets de treinamento e teste
ingredientes = pd.read_csv('./datasets/preprocessed/INGREDIENTES_normalizados_processed.csv')
pratosxingredientes_train = pd.read_csv('./datasets/train/PRATOSxINGREDIENTES_treinamento.csv')


# criando um DataFrame com zeros para os sabores
todos_sabores = [col for col in ingredientes.columns if col not in ['ingrediente', 'categoria']]
sabores_zeros_df = pd.DataFrame(0.0, index=pratosxingredientes_train.index, columns=todos_sabores)

sabores_zeros_df

# concatenando o DataFrame de sabores com o DataFrame de pratos
pratosxingredientes_train = pd.concat([pratosxingredientes_train, sabores_zeros_df], axis=1)

# função para somar os sabores dos ingredientes
def somar_sabores_por_prato(ingredientes_prato):
    soma_sabores = pd.Series(0, index=todos_sabores, dtype=float)
    
    for ingrediente in ingredientes_prato:
        row = ingredientes[ingredientes['ingrediente'] == ingrediente]
        
        if not row.empty:
            valores_numericos = row.iloc[0, 2:].astype(float)  # Ignorar as duas primeiras colunas
            soma_sabores += valores_numericos
    
    return soma_sabores

# aplicando a função de soma dos sabores para cada prato
for index, row in pratosxingredientes_train.iterrows():
    ingredientes_prato = eval(row['ingredientes'])  # Converter a string em uma lista
    sabores_soma = somar_sabores_por_prato(ingredientes_prato)
    
    for sabor in todos_sabores:
        if sabor in pratosxingredientes_train.columns:
            pratosxingredientes_train.loc[index, sabor] = sabores_soma[sabor]

pratosxingredientes_train

pratosxingredientes_train[['ingredientes','pratos_allrecipes','prato','sour', 'bitter', 'sweet', 'fruity', 'grassy','roast', 'caramel', 'fermented','powdery','dry']]

# inicializando o normalizador MinMaxScaler
scaler = MinMaxScaler()

pratosxingredientes_train[todos_sabores] = scaler.fit_transform(pratosxingredientes_train[todos_sabores])

pratosxingredientes_train[['ingredientes','pratos_allrecipes','prato','sour', 'bitter', 'sweet', 'fruity', 'grassy','roast', 'caramel', 'fermented','powdery','dry']]

# exportando base de pratos tratada

pratosxingredientes_train.to_csv('./datasets/preprocessed/PRATOS_processed.csv', index=False)
