==> 6_tratamento_base_final.ipynb <==
import pandas as pd


# Carregar os datasets
cervejas = pd.read_csv('./datasets/preprocessed/CERVEJAS_processed.csv')
harmonizacoes = pd.read_csv('./datasets/preprocessed/HARMONIZACOES_processed.csv')
ingredientes = pd.read_csv('./datasets/preprocessed/INGREDIENTES_normalizados_processed.csv')
pratos = pd.read_csv('./datasets/preprocessed/PRATOS_processed.csv')

harmonizacoes

# transformando o tipo de dado de prato em lista para depois explodi-lo
harmonizacoes['pratos'] = harmonizacoes['pratos'].apply(lambda x: list(eval(x)) if isinstance(x, str) else list(x))

harmonizacoes_exploded

# combinando os dados de cerveja com os dados de pratos para criar o dataset de treino
merge_exploded_cerveja = harmonizacoes_exploded.merge(cervejas, left_on='estilos_cervejas', right_on='estilo')


# fazer o merge baseado no prato individual
merge_exploded_cerveja_pratos = merge_exploded_cerveja.merge(pratos, left_on='pratos', right_on='prato')


# exportando base final com todos os sabores de pratos e atributos de cerveja

merge_exploded_cerveja_pratos.to_csv('./datasets/preprocessed/PRATOxCERVEJA_processed.csv', index=False)

merge_exploded_cerveja_pratos

# Contar a distribuição das classes na coluna 'categoria'
distribuicao_classes2 = merge_exploded_cerveja_pratos['estilos_cervejas'].value_counts()

# Exibir a distribuição
print(distribuicao_classes2)



# Contar a distribuição das classes na coluna 'categoria'
distribuicao_classes3 = merge_exploded_cerveja_pratos['estilos_cervejas_brejas'].value_counts()

# Exibir a distribuição
print(distribuicao_classes3)



# Agrupando pelas colunas e contando o número de ocorrências
contagem = merge_exploded_cerveja.groupby(['estilos_cervejas_brejas', 'pratos', 'category']).size().reset_index(name='count')

# Filtrando para mostrar apenas combinações que aparecem mais de uma vez
contagem_filtrada = contagem[contagem['count'] > 1]

# Ordenando do maior para o menor pela coluna 'count'
contagem_ordenada = contagem_filtrada.sort_values(by='count', ascending=False)

# Exibindo o resultado
print(contagem_ordenada)




# Agrupando por estilos_cervejas_brejas e category e somando as contagens
contagem_final = contagem_filtrada.groupby(['estilos_cervejas_brejas', 'category'])['count'].sum().reset_index()

# Ordenando do maior para o menor pela coluna 'count'
contagem_ordenada = contagem_final.sort_values(by='count', ascending=False)

# Exibindo o resultado
print(contagem_ordenada)


merge_exploded_cerveja_pratos
